+++ src/components/canary/atoms/split-button/index.ts
@@ -1,0 +1,6 @@
+export { SplitButton } from './split-button';
+export type {
+  SplitButtonProps,
+  ArdaSplitButtonStaticConfig,
+  ArdaSplitButtonRuntimeConfig,
+} from './split-button';