+++ src/components/canary/atoms/toggle-group/index.ts
@@ -1,0 +1,8 @@
+export { ToggleGroup, ToggleGroupItem } from './toggle-group';
+export type {
+  ToggleGroupProps,
+  ToggleGroupSingleProps,
+  ToggleGroupMultipleProps,
+  ToggleGroupItemProps,
+  ArdaToggleGroupStaticConfig,
+} from './toggle-group';